% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDiagnosticsFuncs.R
\name{funcResidAnalysis}
\alias{funcResidAnalysis}
\title{Function for running typical residual analysis procedures on models}
\usage{
funcResidAnalysis(modelObject, data, ...)
}
\arguments{
\item{modelObject}{A model object of class lm(), lmer(), nlme(), gls(), nlmer() or others (not tested for others)}

\item{data}{(Optional) A data frame which was used to fit the models. Required only for certain residual plots.}

\item{...}{Optional grouping parameters for testing the grouping of residuals by these parameters.}
}
\value{
A list of plots. Namely * Pearson residuals vs Fitted plot, normality of pearson residuals,
Shapiro-Wilk test of normality of the residuals, qqnorm plot of the residuals,
 boxplots of the residuals grouped by the grouping parameters (if specified).
}
\description{
This function runs the commonly used residual analysis functions on model objects
}
