% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesFuncs.R
\name{funcTableDesign}
\alias{funcTableDesign}
\title{Function for converting a data frame into a beautifully formatted flextable.}
\usage{
funcTableDesign(data_frame, header.name = NULL, tableType = NULL)
}
\arguments{
\item{data_frame}{input data frame.}

\item{header.name}{(Optional) name of the table header. Accepts a string.}

\item{tableType}{(Optional) A string of one of the following * modComp - dataframe with anova comparing models,
*parEst - estimates of the parameters, ideally from the summary table,
 *fixAnova - dataframe with fixed effects anova.}
}
\value{
flextable object
}
\description{
Function for converting a data frame into a beautifully formatted flextable.
}
