% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesFuncs.R
\name{funcTableGrob}
\alias{funcTableGrob}
\title{Function for generating table grob for custom annotation in ggplot2 graphs.}
\usage{
funcTableGrob(data_frame)
}
\arguments{
\item{data_frame}{A data.frame object which needs to be added to a ggplot graph as a tableGrob}
}
\value{
a tableGrob object which can be added to a ggplot2 object using annotation_custom
}
\description{
This function converts a data.frame into a tableGrob object with ttheme_minimal.
}
